Add is_finding_key flag for parsing notes #82
When finding key, the octave number is stripped from the note before parsing to see if there are any errors.
This produced KeyError in a few cases where the note is higher than C (and the key is not C). the is_finding_key flag lets the note be handled differently during parsing.

- removed some debugging statements