An attempt at making Question more generic  #104
I think a question can something else than text, and more, importantly, an answer can definitely be something different than text (an image for instance).
So I tried to build a more generic question class.
For the same reason, I think the questiontype enum should be more on the type of question than on the type of the data. Or we can do both.