improvements to hash and MemoryBrain
The UID can now be created without stamping/sending the Query.

In QueryMemory, added a sort by sent_date, and the possibility to
recall a Query by UID.
The next step would be to add depencies by UID and not only by passing
the object itself.
(Problem remaining: the UID hashing has to be triggered manually, or
the Query be sent.)