changes to init and build_result
Since the result and the UID only depends on input parameters, they are
now set on initialization. This way a Query has an UID from the start
and can be retrieved.

I removed set_result which was confusing (the result cannot be set
externally).
WARNING: the uid is calculated again upon sent(). So if a query is
modified after its creation the UID must be updated. Maybe send() can
return the UID?