progress on #104 and #107
I now try to pass objects in queries for sender and recipient instead
of strings, I hope it will work.
It is necessary for someone to reply to a query to find the object
(=instance of a class) that sent it.
Several other bug fixes (especially in \_\_getattr_\_\) and
improvements (in communication and communicator)