#110 Serious attempt at localization
Borrowed (ok, stolen) code from Alex.
I just have to decide how to get the locale from the user.
On the command line, with os get env, but then where to store it???
Most likely in music maker
On the website, have WebsitePlayer get the info and then pass it to
music maker ?
In the bot cog pass the information to music_maker.